backend/bittan/requirements.txt
django==5.0
psycopg2-binary==2.9.9
django-money==3.4.1
djangorestframework==3.14.0
django-enumfield==3.1
python-dotenv==1.0.1
requests==2.31.0
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
qrcode[pil]==7.4.2
djangorestframework==3.14.0
django-cors-headers==4.3.1